backend/requirements.txt
fastapi==0.68.0
uvicorn==0.15.0
sqlalchemy==1.4.23
psycopg2-binary==2.9.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.5
python-dotenv==0.19.0
pydantic==1.8.2
alembic==1.7.1
tensorflow==2.7.0
pillow==8.3.2
mlflow==1.20.2 